+++ NEWS.md
@@ -5,6 +5,8 @@ This is a summary of all changes since version 1.x.x.
 ## Major changes
 * `missRanger` now also imputes and uses logical variables and character variables.
 
+* Set `imputeSpecial` to `TRUE` to also impute variables of mode numeric but with special classes like time, dates etc. These variables are converted to numeric first, then imputed and in the end set back to their original classes. Currently, the default is `FALSE`. But this might change in later releases.
+
 * Added formula interface to specify which variables to impute (those on the left hand side) and those used to do so (those on the right hand side). Here some (pseudo) examples:
 
   - `. ~ .` (default): Use all variables to impute all variables. Note that only those with missing values will be imputed. Variables without missings will only be used to impute others.
@@ -20,6 +22,8 @@ This is a summary of all changes since version 1.x.x.
 * The first argument of `generateNA` is called `x` instead of `data` in consistency with `imputeUnivariate`.
 
 * `imputeUnivariate` now also works for data frames and matrices.
+
+* In PMM mode, `missRanger` relies on OOB predictions. The smaller the value of `num.trees`, the higher the risk of missing OOB predictions, which caused an error in PMM. Now, `pmm` allows for missing values in `xtrain` or `ytrain`. Thus, the algorithm will even work with `num.trees = 1`. This will be useful to impute large data sets with PMM.
 
 ## Minor changes